Fix: Add explicit integer casts for servo position constraints and update deployment documentation for Raspberry Pi Pico.

+++ src/nerf_standalone/firmware/nerf_micro/Tilt.h
@@ -72,10 +72,10 @@ public:
 
   void setPosition(int us) {
     // Constraint safety
-    if (us < Config::SV_MIN_US)
-      us = Config::SV_MIN_US;
-    if (us > Config::SV_MAX_US)
-      us = Config::SV_MAX_US;
+    if (us < (int)Config::SV_MIN_US)
+      us = (int)Config::SV_MIN_US;
+    if (us > (int)Config::SV_MAX_US)
+      us = (int)Config::SV_MAX_US;
 
     _tiltServo.attach(_pin, Config::SV_MIN_US, Config::SV_MAX_US);
     _tiltServo.writeMicroseconds(us);
